Make git optional
Git will only be used if it is installed and set up with a user.name and
a user.email. Git can be set up at any time. Once it will be set up, the
dream journal will be put under version control on the next dream write.
In addition, all the git operations are now regrouped under the same
function, snapshot.